itemClicked fix
itemClicked is fired on mousedown instead of on click. Moreover, for a lot of use cases, ie. using the sortable as a navigation menu, it is desirable to be able to cancel the click callback when the item is dragged. This is implemented with a flag provided to the itemClicked callback.