Added ability to delay dragStartEvent.
Using angular's $timeout I created a "long click" to start the drag. This allows  for nodes to be clicked on or double clicked without creating buttons that use 'data-nodrag'. It is as simple as adding data-drag-delay to the ui-tree element and type how many milliseconds they want the user to hold the mouse down.