Rewrite horizontal movement calculation.
The increase and decrease of the horizontal level is now based on dragged element
left position.
Each level is specified by a spacing distance (in pixel, customizable as "data-spacing"). A
spacing threshold is computed based on this spacing (spacing / 4).
Now, when the left position of the dragged element is around the parent position +- 
the spacing +- the spacing threshold, it's level is increased/decreased.

You just have to specify the spacing parameter according to your tree style (so that the
visual is approaching the calculation).