Improve vertical placement calculation.
The vertical position is now determined based on the dragged element position
compared to the other element position.
If the top of the dragged element is below next element middle position, then the
dragged element is positionned after it.
If the top of the dragged element is above previous element middle position, then the
dragged element is positionned before it.