Issue #216: rewrite intersection detection
The intersection between dragged element and overlapped element is now
configurable.

Use the parameter "coverage". Can accept an integer between -100 and 100.
Negative number indicates that it's the top border that is used when going
down (and the bottom one when going up). Positive number indicate that it's
the bottom border that is used when going down (and the top one when going
up). And the number is the percentage of coverage of the element before
detecting the overlapping.